% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{bmianthro}
\alias{bmianthro}
\title{BMI Anthro}
\description{
Part of default CDC-derived tables
}
\details{
Contains BMI data for calculating BMI
}
\section{bmianthro.txt.gz}{


Used in function \code{cleangrowth()}
}

