% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_support.R
\name{compute_perc_limit}
\alias{compute_perc_limit}
\title{Compute observation-level percentage criterion limit.
Vectorized over meas.}
\usage{
compute_perc_limit(meas, perclimit_low, perclimit_mid, perclimit_high)
}
\arguments{
\item{meas}{Weight measurements in kg}

\item{perclimit_low}{Limit for wt <= 45 kg}

\item{perclimit_mid}{Limit for 45 < wt <= 80 kg}

\item{perclimit_high}{Limit for wt > 80 kg (0 = disabled: percewma < 0 is never TRUE)}
}
\value{
Numeric vector of perc_limits
}
\description{
Compute observation-level percentage criterion limit.
Vectorized over meas.
}
\keyword{internal}
