% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_support.R
\name{compute_trajectory_fails}
\alias{compute_trajectory_fails}
\title{Pre-computes trajectory rescue for all observations.
Returns TRUE = fails all rescue (not rescued).}
\usage{
compute_trajectory_fails(meas, age_days, err = 5)
}
\description{
Pre-computes trajectory rescue for all observations.
Returns TRUE = fails all rescue (not rescued).
}
\keyword{internal}
