% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Cs}
\alias{Cs}
\title{Character strings from unquoted names}
\usage{
Cs(...)
}
\arguments{
\item{...}{any number of names separated by commas}
}
\value{
character string vector
}
\description{
Makes a vector of character strings from a list of valid S names
}
\keyword{internal}
