% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_support.R
\name{eval_1d}
\alias{eval_1d}
\title{Evaluate single distinct value observations}
\usage{
eval_1d(subj_results, params)
}
\arguments{
\item{subj_results}{data.table with result, meas_m, param, ageyears, id columns
for ONE subject — all params included}

\item{params}{List of limit parameters}
}
\value{
Named character vector of ids to mark as "1D"
}
\description{
Evaluate single distinct value observations
}
\keyword{internal}
