% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_support.R
\name{eval_2d_nonord}
\alias{eval_2d_nonord}
\title{Evaluate 2D non-ordered weight pairs for one subject}
\usage{
eval_2d_nonord(w_subj_df, w_subj_keep, wtallow_formula = "piecewise")
}
\arguments{
\item{w_subj_df}{Weight data with meas_m, age_days, id. Already filtered to Inc.}

\item{w_subj_keep}{Named vector of current exclusion codes for all weight obs}

\item{wtallow_formula}{Formula for wtallow}
}
\value{
Character vector of ids to exclude with "2D Non-Ord" code
}
\description{
Evaluate 2D non-ordered weight pairs for one subject
}
\keyword{internal}
