% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_support.R
\name{eval_error_load}
\alias{eval_error_load}
\title{Evaluate error load for one subject}
\usage{
eval_error_load(subj_results, error_threshold = 0.41)
}
\arguments{
\item{subj_results}{data.table with result, param columns for ONE subject}

\item{error_threshold}{Threshold for error ratio (default 0.41)}
}
\value{
Named character vector of ids to mark as "Error load"
}
\description{
Evaluate error load for one subject
}
\keyword{internal}
