% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_support.R
\name{evil_twins}
\alias{evil_twins}
\title{Evil twins detection for one subject (independent mode)}
\usage{
evil_twins(w_subj_df, wtallow_formula = "piecewise")
}
\arguments{
\item{w_subj_df}{Weight data for one subject (data.table with meas_m, age_days, id, is_rv)
Must include all Inc values (including RVs in independent mode)}

\item{wtallow_formula}{Formula for wtallow/ET caps}
}
\value{
Character vector of ids to exclude with "Evil twins" code
}
\description{
Evil twins detection for one subject (independent mode)
}
\keyword{internal}
