% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{fent_for_corraga}
\alias{fent_for_corraga}
\title{Fenton-Based Corrected Age Reference (Weight to GA Lookup)}
\description{
Reference values for premature infant corrected-age estimation, derived from
data extracted from published Fenton (2013) growth curves. Provides a
reverse lookup from integer weight (grams, rounded to 10) to estimated
gestational age in days, by sex.
}
\section{fent_for_corraga.csv.gz}{


Used in function \code{cleangrowth()}
}

