% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{fent_for_corrz}
\alias{fent_for_corrz}
\title{Fenton-Based Corrected Z-Score Reference}
\description{
Reference values for premature infant corrected z-score calculation, derived
from data extracted from published Fenton (2013) growth curves. Provides
weight, length, and head circumference reference values by gestational age
and sex.
}
\section{fent_for_corrz.csv.gz}{


Used in function \code{cleangrowth()}
}

