% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{fent_foraga}
\alias{fent_foraga}
\title{Fenton Weight-to-GA Lookup}
\description{
Fenton growth curves with premature infant data mapping sex and integer
weight (grams, rounded to nearest 10) to estimated gestational age in days.
}
\section{fent_foraga.csv.gz}{


Used in function \code{cleangrowth()} for GA correction (Step 2b).
}

