% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{fentlms_foraga}
\alias{fentlms_foraga}
\title{Fenton Growth Curves}
\description{
Fenton growth curves with premature infant data with sex, age, and integer
weight
}
\section{fentlms_foraga.csv.gz}{


Used in function \code{cleangrowth()}
}

