% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{fentlms_forz}
\alias{fentlms_forz}
\title{Fenton Growth Curve Z-Scores}
\description{
Fenton growth curves with premature infant z-scores for height and head
circumference
}
\section{fentlms_forz.csv.gz}{


Used in function \code{cleangrowth()}
}

