% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{fenton2025_ms_lookup_smoothed}
\alias{fenton2025_ms_lookup_smoothed}
\title{Fenton 2025 Growth Reference (M, S_upper, S_lower)}
\description{
Fenton 2025 growth reference for premature infants. Contains median (M) and
upper/lower standard deviations (S_upper, S_lower) extracted from plotted
Fenton 2025 curves, for weight, length, and head circumference by sex and
gestational age in days. Used with the CSD method for z-score calculation.
}
\section{fenton2025_ms_lookup_smoothed.csv}{


Used in function \code{cleangrowth()} for GA correction (Step 2b).
}

