% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{growth_cdc_ext}
\alias{growth_cdc_ext}
\title{CDC Growth Percentile Table}
\description{
Part of default CDC-derived tables
}
\details{
Contains percentiles for various ages, gender, and weights, pre-calculated by CDC
}
\section{growthfile_cdc_ext.csv.gz}{


Used in function \code{cleangrowth()}
}

