% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{growth_cdc_ext_infants}
\alias{growth_cdc_ext_infants}
\title{CDC Growth Percentile Table for Infants}
\description{
Part of default CDC-derived tables
}
\details{
Contains percentiles for various ages, gender, and weights, pre-calculated by CDC for infants algorithm
}
\section{growthfile_cdc_ext_infants.csv.gz}{


Used in function \code{cleangrowth()}
}

