% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{growth_who_ext}
\alias{growth_who_ext}
\title{WHO Growth Percentile Table}
\description{
Part of default WHO-derived tables
}
\details{
Contains percentiles for various ages, gender, and weights, pre-calculated by WHO
}
\section{growthfile_who.csv.gz}{


Used in function \code{cleangrowth()}
}

