% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{lenanthro}
\alias{lenanthro}
\title{Length to Age Table}
\description{
Part of default CDC-derived tables
}
\details{
Contains percentiles for various ages, gender, and weights, pre-calculated by CDC
}
\section{lenanthro.txt}{


Used in function \code{cleangrowth()}
}

