% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{nhanes-reference-medians}
\alias{nhanes-reference-medians}
\title{NHANES reference medians}
\description{
Contains reference median values for default recentering, derived from NHANES
}
\section{nhanes-reference-medians.csv}{


Used in function \code{cleangrowth()}
}

