% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{rc-reference-medians}
\alias{rc-reference-medians}
\title{Infants reference medians}
\description{
Contains reference median values for default recentering in the infants
algorithm
}
\section{rcfile-2023-08-15_format.csv.gz}{


Used in function \code{cleangrowth()}
}

