% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{read.anthro}
\alias{read.anthro}
\title{Read anthro}
\usage{
read.anthro(path = "", cdc.only = F)
}
\value{
Function for calculating BMI based on measurement, age in days, sex, and measurement value.
}
\description{
Read anthro
}
