% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{read_anthro}
\alias{read_anthro}
\title{Read anthro}
\usage{
read_anthro(path = "", cdc.only = F)
}
\value{
Function for calculating BMI based on measurement, age in days, sex, and measurement value.
}
\description{
Read anthro
}
\examples{
# Return calculating function with all defaults
afunc <- read_anthro()

# Return calculating function while specifying a path and using only CDC data
afunc <- read_anthro(path = system.file("extdata", package = "growthcleanr"),
                     cdc.only = TRUE)
}
