% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{sd.median}
\alias{sd.median}
\title{Calculate median SD score by age for each parameter.}
\usage{
sd.median(param, sex, agedays, sd.orig)
}
\arguments{
\item{param}{Vector identifying each measurement, may be 'WEIGHTKG', or 'HEIGHTCM'.}

\item{sex}{Vector identifying the gender of the subject, may be 'M', 'm', or 0 for males, vs. 'F', 'f' or 1 for females.}

\item{agedays}{Numeric vector containing the age in days at each measurement.}

\item{sd.orig}{Vector of previously calculated standard deviation (SD) scores for each measurement before re-centering.}
}
\value{
Table of data with median SD-scores per day of life by gender and parameter.
}
\description{
Calculate median SD score by age for each parameter.
}
