% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitinput}
\alias{splitinput}
\title{splitinput}
\usage{
splitinput(
  df,
  fname = deparse(substitute(df)),
  fdir = "",
  min_nrow = 10000,
  keepcol = "subjid"
)
}
\arguments{
\item{df}{data frame to split}

\item{fname}{new name for each of the split files to start with}

\item{fdir}{directory to put each of the split files (default working directory)}

\item{keepcol}{the column name (default "subjid") to use to keep records with the same values together in the same single split file}

\item{min_row}{minimum number of rows for each split file (default 10000)}
}
\value{
the count number refering to the last split file written
}
\description{
\code{splitinput} Splits input based on keepcol specified, yielding csv files each with at least the mininum
number of rows that are written and saved separately (except for the last split file written, which may be
smaller). Allows splitting input data while ensuring all records for each individual subject will stay together
in one file. Pads split filenames with zeros out to five digits for consistency, assuming < 100,000 file count
result.
}
