% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{tanner_ht_vel}
\alias{tanner_ht_vel}
\title{Tanner Growth Velocity Table}
\description{
Part of default CDC-derived tables
}
\details{
Contains velocities for growth pre-calculated by CDC
}
\section{tanner_ht_vel.csv.gz}{


Used in function \code{cleangrowth()}
}

