% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{tanner_ht_vel_2sd}
\alias{tanner_ht_vel_2sd}
\title{Tanner Growth Velocity Table with (2\eqn{\sigma})}
\description{
Part of default CDC-derived tables
}
\details{
Contains velocities for growth pre-calculated by CDC, including those 2
standard deviations away.
}
\section{tanner_ht_vel.csv}{


Used in function \code{acf_answers()}
}

