% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{tanner_ht_vel_rev}
\alias{tanner_ht_vel_rev}
\title{Tanner Growth Velocity Table for Infants}
\description{
Part of default CDC-derived tables
}
\details{
Contains velocities for growth pre-calculated by CDC, used for the infants
algorithm
}
\section{tanner_ht_vel_rev.csv.gz}{


Used in function \code{cleangrowth()}
}

