% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{test_syngrowth_sas_output_compare}
\alias{test_syngrowth_sas_output_compare}
\title{CDC SAS BMI Output}
\description{
Contains results of CDC SAS macro for calculating BMI values.
}
\section{test_syngrowth_sas_output_compare.csv.gz}{


Used to test function \code{ext_bmiz()}
}

