% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{test_syngrowth_wide}
\alias{test_syngrowth_wide}
\title{CDC SAS BMI Input}
\description{
Contains input data for CDC SAS macro for calculating BMI values.
}
\section{test_syngrowth_wide.csv.gz}{


Used to test function \code{ext_bmiz()}
}

