% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{weianthro}
\alias{weianthro}
\title{Weight Anthro Table}
\description{
Part of default CDC-derived tables
}
\details{
Contains median and standard deviation for weight by age and gender
}
\section{weianthro.csv}{


Used in function \code{cleangrowth()}
}

