% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{who_hc_maxvel}
\alias{who_hc_maxvel}
\title{WHO Maximum Head Circumference Velocity for (3\eqn{\sigma})}
\description{
Part of default WHO-derived tables
}
\details{
Contains three standard deviations for the World Health Organization values of
maximum head circumference velocities.
}
\section{who_hc_maxvel_3sd_infants.csv.gz}{


Used in function \code{cleangrowth()}
}

