% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{who_hc_vel_3sd}
\alias{who_hc_vel_3sd}
\title{WHO Head Circumference Velocity for (3\eqn{\sigma})}
\description{
Part of default WHO-derived tables
}
\details{
Contains three standard deviations for the World Health Organization values of
head circumference velocities.
}
\section{who_hc_vel_3sd_infants.csv.gz}{


Used in function \code{cleangrowth()}
}

