% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{who_ht_maxvel}
\alias{who_ht_maxvel}
\title{WHO Maximum Height Velocity for (3\eqn{\sigma})}
\description{
Part of default WHO-derived tables
}
\details{
Contains three standard deviations for the World Health Organization values of
maximum height velocities.
}
\section{who_ht_maxvel_3sd.csv.gz}{


Used in function \code{cleangrowth()}
}

