% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{who_ht_vel_2sd}
\alias{who_ht_vel_2sd}
\title{WHO Height Velocity for (2\eqn{\sigma})}
\description{
Part of default WHO-derived tables
}
\details{
Contains two standard deviations for the World Health Organization values of
height velocities.
}
\section{who_ht_vel_2sd.csv.gz}{


Used in function \code{acf_answers()}
}

