% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{who_ht_vel_3sd}
\alias{who_ht_vel_3sd}
\title{WHO Height Velocity for (3\eqn{\sigma})}
\description{
Part of default WHO-derived tables
}
\details{
Contains three standard deviations for the World Health Organization values of
height velocities.
}
\section{who_ht_vel_3sd.csv}{


Used in function \code{cleangrowth()}
}

