% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustcarryforward.R
\name{adjustcarryforward}
\alias{adjustcarryforward}
\title{adjustcarryforward
\code{adjustcarryforward} Uses absolute height velocity to identify values
excluded as carried forward values for reinclusion.}
\usage{
adjustcarryforward(
  subjid,
  param,
  agedays,
  sex,
  measurement,
  orig.exclude,
  sd.recenter = NA,
  ewma.exp = -1.5,
  ref.data.path = "",
  quietly = T,
  minfactor = 0.5,
  maxfactor = 2,
  banddiff = 3,
  banddiff_plus = 5.5,
  min_ht.exp_under = 2,
  min_ht.exp_over = 0,
  max_ht.exp_under = 0.33,
  max_ht.exp_over = 1.5
)
}
\arguments{
\item{max_ht.exp_over}{}
}
\value{

}
\description{
adjustcarryforward
\code{adjustcarryforward} Uses absolute height velocity to identify values
excluded as carried forward values for reinclusion.
}
