% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{cleangrowth}
\alias{cleangrowth}
\title{Clean growth measurements}
\usage{
cleangrowth(subjid, param, agedays, sex, measurement, recover.unit.error = F,
  sd.extreme = 25, z.extreme = 25, height.tolerance.cm = 2.5,
  too.many.errors = 3, sd.recenter = NA, include.carryforward = F,
  ewma.exp = -1.5, ref.data.path = "", log.path = NA, parallel = F,
  num.batches = NA, quietly = T)
}
\arguments{
\item{subjid}{Vector of unique identifiers for each subject in the database.}

\item{param}{Vector identifying each measurement, may be 'WEIGHTKG', 'HEIGHTCM', or 'LENGTHCM'
'HEIGHTCM' vs. 'LENGTHCM' only affects z-score calculations between ages 24 to 35 months (730 to 1095 days).
All linear measurements below 731 days of life (age 0-23 months) are interpreted as supine length, and
all linear measurements above 1095 days of life (age 36+ months) are interpreted as standing height.}

\item{agedays}{Numeric vector containing the age in days at each measurement.}

\item{sex}{Vector identifying the gender of the subject, may be 'M', 'm', or 0 for males, vs. 'F', 'f' or 1 for females.}

\item{measurement}{Numeric vector containing the actual measurement data.  Weight must be in
kilograms (kg), and linear measurements (height vs. length) in centimeters (cm).}

\item{recover.unit.error}{Indicates whether the cleaning algorithm should
attempt to identify unit errors (I.e. inches vs. cm, lbs vs. kg). If unit
errors are identified, the value will be corrected and retained within the
cleaning algorithm as a valid measurement.  Defaults to FALSE.}

\item{sd.extreme}{Measurements more than sd.extreme standard deviations from
the mean (either above or below) will be flagged as invalid. Defaults to 25.}

\item{z.extreme}{Measurements with an absolute z-score greater than
z.extreme will be flagged as invalid. Defaults to 25.}

\item{include.carryforward}{Determines whether Carry-Forward values are kept in the output.  Defaults to False.}

\item{ewma.exp}{Exponent to use for weighting measurements in the
exponentially weighted moving average calculations. Defaults to -1.5.
This exponent should be negative in order to weight growth measurements
closer to the measurement being evaluated more strongly. Exponents that are
further from zero (e.g. -3) will increase the relative influence of
measurements close in time to the measurement being evaluated compared to
using the default exponent.}

\item{ref.data.path}{Path to reference data. If not supplied, the year 2000
Centers for Disease Control (CDC) reference data will be used.}

\item{log.path}{Path to log file output when running in parallel (non-quiet mode).}

\item{parallel}{Determines if function runs in parallel.  Defaults to FALSE.}

\item{num.batches}{Specify the number of batches to run in parallel. Only
applies if parallel is set to TRUE. Defaults to the number of workers
returned by the getDoParWorkers function in the foreach package.}

\item{quietly}{Determines if function messages are to be displayed and if log files (parallel only) are to be generated.
Defaults to TRUE}

\item{sd.receiver}{Data frame or table with median SD-scores per day of life
by gender and parameter. Columns in the table must include param, sex,
agedays, and sd.median.  If not supplied, the median values will be
calculated using the growth data that is being cleaned. Defaults to NA.}
}
\value{
Vector of exclusion codes for each of the input measurements.

Possible values for each code are:
\itemize{
\item 'Include', 'Unit-Error-High', 'Unit-Error-Low', 'Swapped-Measurements', 'Missing',
\item 'Exclude-Carried-Forward', 'Exclude-SD-Cutoff', 'Exclude-EWMA-Extreme', 'Exclude-EWMA-Extreme-Pair',
\item 'Exclude-Duplicate', 'Exclude-EWMA-8', 'Exclude-EWMA-9', 'Exclude-EWMA-10', 'Exclude-EWMA-11', 'Exclude-EWMA-12', 'Exclude-EWMA-13', 'Exclude-EWMA-14',
\item 'Exclude-Min-Height-Change', 'Exclude-Max-Height-Change',
\item 'Exclude-Pair-Delta-17', 'Exclude-Pair-Delta-18', 'Exclude-Single-Outlier', 'Exclude-Too-Many-Errors', 'Exclude-Too-Many-Errors-Other-Parameter'
}
}
\description{
Clean growth measurements
}
\examples{
# Standard calculation Using Provided Age/Weight Data
colnames(data_stats)<-c("id", "measure", "agedays", "sex", "value")
clean_stats<-cleangrowth(data_stats$id, data_stats$measure, data_stats$agedays, data_stats$sex, data_stats$value)

# Once processed you can filter data based on result value
data_stats<-data_stats[,clean_result:=clean_stats]
clean_data_stats<-data_stats[clean_result="Include"] # only include data marked as "Include"

# Parallel Processing (currently only supported on Unix/Linux)
# Run using 6 cores and batches
clean_stats<-cleangrowth(data_stats$id, data_stats$measure, data_stats$agedays, data_stats$sex, data_stats$value, parallel = TRUE, num.batches = 6)

}
