% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{cleangrowth}
\alias{cleangrowth}
\title{Clean growth measurements}
\usage{
cleangrowth(
  subjid,
  param,
  agedays,
  sex,
  measurement,
  recover.unit.error = FALSE,
  sd.extreme = 25,
  z.extreme = 25,
  lt3.exclude.mode = "default",
  height.tolerance.cm = 2.5,
  error.load.mincount = 2,
  error.load.threshold = 0.5,
  sd.recenter = NA,
  sdmedian.filename = "",
  sdrecentered.filename = "",
  include.carryforward = FALSE,
  ewma.exp = -1.5,
  ref.data.path = "",
  log.path = ".",
  parallel = FALSE,
  num.batches = NA,
  quietly = TRUE
)
}
\arguments{
\item{subjid}{Vector of unique identifiers for each subject in the database.}

\item{param}{Vector identifying each measurement, may be 'WEIGHTKG', 'HEIGHTCM', or 'LENGTHCM'
'HEIGHTCM' vs. 'LENGTHCM' only affects z-score calculations between ages 24 to 35 months (730 to 1095 days).
All linear measurements below 731 days of life (age 0-23 months) are interpreted as supine length, and
all linear measurements above 1095 days of life (age 36+ months) are interpreted as standing height.}

\item{agedays}{Numeric vector containing the age in days at each measurement.}

\item{sex}{Vector identifying the gender of the subject, may be 'M', 'm', or 0 for males, vs. 'F', 'f' or 1 for females.}

\item{measurement}{Numeric vector containing the actual measurement data.  Weight must be in
kilograms (kg), and linear measurements (height vs. length) in centimeters (cm).}

\item{recover.unit.error}{Indicates whether the cleaning algorithm should
attempt to identify unit errors (I.e. inches vs. cm, lbs vs. kg). If unit
errors are identified, the value will be corrected and retained within the
cleaning algorithm as a valid measurement.  Defaults to FALSE.}

\item{sd.extreme}{Measurements more than sd.extreme standard deviations from
the mean (either above or below) will be flagged as invalid. Defaults to 25.}

\item{z.extreme}{Measurements with an absolute z-score greater than
z.extreme will be flagged as invalid. Defaults to 25.}

\item{lt3.exclude.mode}{Determines type of exclusion procedure to use for 1 or 2 measurements of one type without
matching same ageday measurements for the other parameter. Options include "default" (standard growthcleanr approach),
and "flag.both" (in case of two measurements of one type without matching values for the other parameter, flag both
for exclusion if beyond threshold)}

\item{error.load.mincount}{minimum count of exclusions on parameter before
considering excluding all measurements. Defaults to 2.}

\item{error.load.threshold}{threshold of percentage of excluded measurement count to included measurement
count that must be exceeded before excluding all measurements of either parameter. Defaults to 0.5.}

\item{sd.recenter}{Data frame or table with median SD-scores per day of life
by gender and parameter. Columns in the table must include param, sex,
agedays, and sd.median.  If not supplied, the median values will be
calculated using the growth data that is being cleaned. Defaults to NA.}

\item{sdmedian.filename}{Name of file to save sd.median data calculated on the input dataset to as CSV.
Defaults to "", for which this data will not be saved. Use for extracting medians for parallel processing
scenarios other than the built-in parallel option.}

\item{sdrecentered.filename}{Name of file to save re-centered data to as CSV. Defaults to "", for which this
data will not be saved. Useful for post-processing and debugging.}

\item{include.carryforward}{Determines whether Carry-Forward values are kept in the output. Defaults to False.}

\item{ewma.exp}{Exponent to use for weighting measurements in the
exponentially weighted moving average calculations. Defaults to -1.5.
This exponent should be negative in order to weight growth measurements
closer to the measurement being evaluated more strongly. Exponents that are
further from zero (e.g. -3) will increase the relative influence of
measurements close in time to the measurement being evaluated compared to
using the default exponent.}

\item{ref.data.path}{Path to reference data. If not supplied, the year 2000
Centers for Disease Control (CDC) reference data will be used.}

\item{log.path}{Path to log file output when running in parallel (non-quiet mode). Default is ".". A new
directory will be created if necessary.}

\item{parallel}{Determines if function runs in parallel.  Defaults to FALSE.}

\item{num.batches}{Specify the number of batches to run in parallel. Only
applies if parallel is set to TRUE. Defaults to the number of workers
returned by the getDoParWorkers function in the foreach package.}

\item{quietly}{Determines if function messages are to be displayed and if log files (parallel only) are to be generated.
Defaults to TRUE}
}
\value{
Vector of exclusion codes for each of the input measurements.

Possible values for each code are:
\itemize{
\item 'Include', 'Unit-Error-High', 'Unit-Error-Low', 'Swapped-Measurements', 'Missing',
\item 'Exclude-Carried-Forward', 'Exclude-SD-Cutoff', 'Exclude-EWMA-Extreme', 'Exclude-EWMA-Extreme-Pair',
\item 'Exclude-Duplicate',
\item 'Exclude-EWMA-8', 'Exclude-EWMA-9', 'Exclude-EWMA-10', 'Exclude-EWMA-11', 'Exclude-EWMA-12', 'Exclude-EWMA-13', 'Exclude-EWMA-14',
\item 'Exclude-Min-Height-Change', 'Exclude-Max-Height-Change',
\item 'Exclude-Pair-Delta-17', 'Exclude-Pair-Delta-18', 'Exclude-Pair-Delta-19',
\item 'Exclude-Single-Outlier', 'Exclude-Too-Many-Errors', 'Exclude-Too-Many-Errors-Other-Parameter'
}
}
\description{
Clean growth measurements
}
\examples{
# Run calculation using a small subset of given data
df_stats <- as.data.frame(syngrowth)
df_stats <- df_stats[df_stats$subjid \%in\% unique(df_stats[, "subjid"])[1:5], ]

clean_stats <- cleangrowth(
  subjid = df_stats$subjid,
  param = df_stats$param,
  agedays = df_stats$agedays,
  sex = df_stats$sex,
  measurement = df_stats$measurement
)

# Once processed you can filter data based on result value
df_stats <- cbind(df_stats, "clean_result" == clean_stats)
clean_df_stats <- df_stats[, df_stats$clean_result == "Include"]

# Parallel processing: run using 3 cores and batches
df_stats <- cleangrowth(
  subjid = df_stats$subjid,
  param = df_stats$param,
  agedays = df_stats$agedays,
  sex = df_stats$sex,
  measurement = df_stats$measurement,
  parallel = TRUE,
  num.batches = 2
)
}
