% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{ewma}
\alias{ewma}
\title{Exponentially Weighted Moving Average (EWMA)}
\usage{
ewma(agedays, z, ewma.exp, ewma.adjacent = T)
}
\arguments{
\item{agedays}{Vector of age in days for each z score (potentially transformed to adjust weighting).}

\item{z}{Input vector of numeric z-score data.}

\item{ewma.exp}{Exponent to use for weighting.}

\item{ewma.adjacent}{Specify whether EWMA values excluding adjacent measurements should be calculated.  Defaults to TRUE.}
}
\value{
Data frame with 3 variables:
\itemize{
\item The first variable (ewma.all) contains the EWMA at observation time
excluding only the actual observation for that time point.
\item The second variable (ewma.before) contains the EWMA for each observation excluding both the actual observation
and the immediate prior observation.
\item The third variable (ewma.after) contains the EWMA for each observation excluding both the actual observation
and the subsequent observation.
}
}
\description{
\code{ewma} calculates the exponentially weighted moving average (EWMA) for a set of numeric observations over time.
}
