% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{read.anthro}
\alias{read.anthro}
\title{Function to calculate z-scores and csd-scores based on anthro tables.}
\usage{
read.anthro(path = "", cdc.only = F)
}
\arguments{
\item{path}{Path to supplied reference anthro data. Defaults to package anthro tables.}

\item{cdc.only}{Whether or not only CDC data should be used. Defaults to false.}
}
\value{
Function for calculating BMI based on measurement, age in days, sex, and measurement value.
}
\description{
Function to calculate z-scores and csd-scores based on anthro tables.
}
