% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syngrowth}
\alias{syngrowth}
\title{syngrowth}
\format{
A data frame with six variables: \code{id}, \code{subjid},
\code{sex}, \code{agedays}, \code{param}, and \code{measurement}
}
\usage{
syngrowth
}
\description{
A synthetic set of growth measurements from ~2,500
subjects generated using Synthea, with measurement errors
for testing with growthcleanr.
}
\details{
Example pediatric growth (heightcm, weightkg) data.
}
\keyword{datasets}
